% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gq_normal}
\alias{gq_normal}
\title{Gaussian quadrature of normal density}
\usage{
gq_normal(n)
}
\arguments{
\item{n}{number of nodes}
}
\description{
Return node values and weights from Gaussian quadrature of
  normal distribution with n nodes.
}
