% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{human_equilibrium}
\alias{human_equilibrium}
\title{Equilibrium solution}
\usage{
human_equilibrium(EIR, ft, p, age, h = gq_normal(10))
}
\arguments{
\item{EIR}{EIR for adults, in units of infectious bites per person per year}

\item{ft}{proportion of clinical cases effectively treated}

\item{p}{vector of model parameters}

\item{age}{vector of age groups, in units of years}

\item{h}{a list of Gauss-Hermite nodes and associated weights, used for
integrating over heterogeneity in biting. See \code{?gq_normal} for an
example.}
}
\description{
Returns the equilibrium states for the model of Griffin et al.
  (2014). A derivation of the equilibrium solutions can be found in Griffin
  (2016). Integrates over the distribution of biting heterogeneity using
  Gaussian quadrature.
}
\references{
Griffin et. al. (2014). Estimates of the changing age-burden of
  Plasmodium falciparum malaria disease in sub-Saharan Africa.
  doi:10.1038/ncomms4136

  Griffin (2016). Is a reproduction number of one a threshold for Plasmodium
  falciparum malaria elimination? doi:10.1186/s12936-016-1437-9 (see
  supplementary material)
}
