% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_parameter_set}
\alias{load_parameter_set}
\title{Laod a Parameter Set from File}
\usage{
load_parameter_set(file_name = "Jamie_parameters.rds")
}
\arguments{
\item{file_name}{the name of a parameter set within the
inst/extdata/parameter_sets folder.}
}
\description{
Parameter sets are stored within the package
  inst/extdata/parameter_sets folder. Load one of these sets by name.
}
