% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{malariaEq_file}
\alias{malariaEq_file}
\title{Load sytem file for this package}
\usage{
malariaEq_file(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Load and return file from within the inst folder of this
  package.
}
