% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaEquilibrium.R
\docType{package}
\name{malariaEquilibrium}
\alias{malariaEquilibrium}
\title{malariaEquilibrium}
\description{
Often we are interested in the state of a given malaria
  transmission model at equibrium. However, some models (e.g. the Griffin et
  al. 2014 model) are quite complex, and can result in equilibrium solutions
  that are fairly in-depth. In these situations it is useful to have a
  "canonical" equilibrium solution that is tried and tested, and can be used
  reliably by multiple users. This package aims to be a place for hosting
  these canonical solutions, and for storing useful tests, checks and
  plotting functions for exploring a given solution.
}
