% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{square}
\alias{square}
\title{Square a vector of values}
\usage{
square(x = 1:5)
}
\arguments{
\item{x}{vector of values.}
}
\description{
Simple test function that demonstrates some of the features of
  this package by squaring an input vector of values.
}
\examples{
# Find square of first 100 values
square(1:100)
}
